requirements.txt
moviepy==1.0.3
librosa==0.10.1
SpeechRecognition==3.10.0
pydub==0.25.1
ffmpeg-python==0.2.0


